task1/requirements.txt
praw
pytrends
requests
newsapi-python
vaderSentiment
pandas
python-dotenv
